datamanagers now working for connections.
