+++ rdfframework/datamanager/vocabularies/adms.ttl
