+++ rdfframework/datamanager/vocabularies/foaf.rdf
@@ -4,23 +4,22 @@
 <!--   FOAF specification, http://xmlns.com/foaf/spec/                             -->
 <!--                                                                             -->
 <!-- first we introduce a number of RDF namespaces we will be using... -->
-<rdf:RDF 
-	xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" 
-	xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#" 
-	xmlns:owl="http://www.w3.org/2002/07/owl#" 
-	xmlns:vs="http://www.w3.org/2003/06/sw-vocab-status/ns#" 
-	xmlns:foaf="http://xmlns.com/foaf/0.1/" 
-	xmlns:wot="http://xmlns.com/wot/0.1/" 
+<rdf:RDF
+	xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+	xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
+	xmlns:owl="http://www.w3.org/2002/07/owl#"
+	xmlns:vs="http://www.w3.org/2003/06/sw-vocab-status/ns#"
+	xmlns:foaf="http://xmlns.com/foaf/0.1/"
+	xmlns:wot="http://xmlns.com/wot/0.1/"
 	xmlns:dc="http://purl.org/dc/elements/1.1/">
 <!-- Here we describe general characteristics of the FOAF vocabulary ('ontology'). -->
   <owl:Ontology rdf:about="http://xmlns.com/foaf/0.1/" dc:title="Friend of a Friend (FOAF) vocabulary" dc:description="The Friend of a Friend (FOAF) RDF vocabulary, described using W3C RDF Schema and the Web Ontology Language." >
   </owl:Ontology>
 
-
   <!-- OWL/RDF interop section - geeks only -->
   <!--  most folk can ignore this lot. the game here is to make FOAF
-  	work with vanilla RDF/RDFS tools, and with the stricter OWL DL 
-	profile of OWL. At the moment we're in the OWL Full flavour of OWL. 
+  	work with vanilla RDF/RDFS tools, and with the stricter OWL DL
+	profile of OWL. At the moment we're in the OWL Full flavour of OWL.
 	The following are tricks to try have the spec live in both worlds
 	at once. See
 		http://phoebus.cs.man.ac.uk:9999/OWL/Validator
@@ -29,7 +28,7 @@
   <owl:AnnotationProperty rdf:about="http://xmlns.com/wot/0.1/assurance"/>
   <owl:AnnotationProperty rdf:about="http://xmlns.com/wot/0.1/src_assurance"/>
   <owl:AnnotationProperty rdf:about="http://www.w3.org/2003/06/sw-vocab-status/ns#term_status"/>
-  <!--  DC terms are NOT annotation properties in general, so we consider the following 
+  <!--  DC terms are NOT annotation properties in general, so we consider the following
 	claims scoped to this document. They may be removed in future revisions if
 	OWL tools become more flexible. -->
   <owl:AnnotationProperty rdf:about="http://purl.org/dc/elements/1.1/description"/>
@@ -59,8 +58,8 @@
     <rdfs:subClassOf><owl:Class rdf:about="http://xmlns.com/foaf/0.1/Agent"/></rdfs:subClassOf>
 <!--    <rdfs:subClassOf><owl:Class rdf:about="http://xmlns.com/wordnet/1.6/Agent"/></rdfs:subClassOf> -->
     <rdfs:subClassOf><owl:Class rdf:about="http://www.w3.org/2003/01/geo/wgs84_pos#SpatialThing" rdfs:label="Spatial Thing"/></rdfs:subClassOf>
-    <!-- aside: 
-	are spatial things always spatially located? 
+    <!-- aside:
+	are spatial things always spatially located?
 	Person includes imaginary people... discuss... -->
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
 
@@ -114,7 +113,7 @@
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#Class"/>
     <rdfs:subClassOf rdf:resource="http://xmlns.com/foaf/0.1/Document"/>
   </rdfs:Class>
-	
+
   <rdfs:Class rdf:about="http://xmlns.com/foaf/0.1/OnlineAccount" vs:term_status="testing" rdfs:label="Online Account" rdfs:comment="An online account.">
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#Class"/>
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
@@ -136,7 +135,7 @@
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
   </rdfs:Class>
 <!-- FOAF properties (ie. relationships). -->
-  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/mbox" vs:term_status="stable" rdfs:label="personal mailbox" rdfs:comment="A 
+  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/mbox" vs:term_status="stable" rdfs:label="personal mailbox" rdfs:comment="A
 personal mailbox, ie. an Internet mailbox associated with exactly one owner, the first owner of this mailbox. This is a 'static inverse functional property', in that  there is (across time and change) at most one individual that ever has any particular value for foaf:mbox.">
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#InverseFunctionalProperty"/>
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#ObjectProperty"/>
@@ -148,15 +147,15 @@ personal mailbox, ie. an Internet mailbox associated with exactly one owner, the
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#InverseFunctionalProperty"/>
 
 <!--
-put it back in again 2006-01-29 - see 
+put it back in again 2006-01-29 - see
 http://chatlogs.planetrdf.com/swig/2006-01-29.html#T21-12-35
 I have mailed rdfweb-dev@vapours.rdfweb.org for discussion.
 Libby
 
 Commenting out as a kindness to OWL DL users. The semantics didn't quite cover
-our usage anyway, since (a) we want static-across-time, which is so beyond OWL as 
-to be from another planet (b) we want identity reasoning invariant across xml:lang 
-tagging. FOAF code will know what to do. OWL folks note, this assertion might return. 
+our usage anyway, since (a) we want static-across-time, which is so beyond OWL as
+to be from another planet (b) we want identity reasoning invariant across xml:lang
+tagging. FOAF code will know what to do. OWL folks note, this assertion might return.
 
 danbri
 -->
@@ -166,15 +165,15 @@ danbri
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
   </rdf:Property>
 
-  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/gender" vs:term_status="testing" 
-rdfs:label="gender" 
+  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/gender" vs:term_status="testing"
+rdfs:label="gender"
 rdfs:comment="The gender of this Agent (typically but not necessarily 'male' or 'female').">
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#FunctionalProperty"/>
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#DatatypeProperty"/>
     <rdfs:domain rdf:resource="http://xmlns.com/foaf/0.1/Agent"/>
     <rdfs:range rdf:resource="http://www.w3.org/2000/01/rdf-schema#Literal"/>
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
-    <!-- whatever one's gender is, and we are liberal in leaving room for more options 
+    <!-- whatever one's gender is, and we are liberal in leaving room for more options
     than 'male' and 'female', we model this so that an agent has only one gender. -->
   </rdf:Property>
 
@@ -220,7 +219,7 @@ rdfs:comment="The gender of this Agent (typically but not necessarily 'male' or 
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
 <!--
     <rdfs:subPropertyOf rdf:resource="http://xmlns.com/foaf/0.1/nick"/>
-...different to the other IM IDs, as Jabber has wider usage, so 
+...different to the other IM IDs, as Jabber has wider usage, so
 we don't want the implied rdfs:domain here.
 
 -->
@@ -376,7 +375,7 @@ we don't want the implied rdfs:domain here.
     <rdfs:isDefinedBy rdf:resource="http://xmlns.com/foaf/0.1/"/>
     <owl:inverseOf rdf:resource="http://xmlns.com/foaf/0.1/maker"/>
   </rdf:Property>
-  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/maker"  vs:term_status="stable" rdfs:label="maker" rdfs:comment="An agent that 
+  <rdf:Property rdf:about="http://xmlns.com/foaf/0.1/maker"  vs:term_status="stable" rdfs:label="maker" rdfs:comment="An agent that
 made this thing.">
     <owl:equivalentProperty rdf:resource="http://purl.org/dc/terms/creator"/>
     <rdf:type rdf:resource="http://www.w3.org/2002/07/owl#ObjectProperty"/>
