+++ rdfframework/datamanager/vocabularies/owl.ttl
