+++ rdfframework/datamanager/vocabularies/rdf.ttl
