+++ rdfframework/datamanager/vocabularies/rdfs.ttl
