+++ rdfframework/datamanager/vocabularies/schema.nt
