+++ rdfframework/datamanager/vocabularies/vcard.ttl
