+++ rdfframework/datamanager/vocabularies/void.ttl
