+++ rdfframework/utilities/__init__.py
@@ -8,5 +8,5 @@ from .baseutilities import *
 from .frameworkutilities import *
 from .valuecalculator import *
 from .formattingfunctions import *
-from .fileutilities import list_files
+from .fileutilities import list_files, is_writeable_dir
 from .gitutilities import *
