+++ tests/connections/test_blazegraph.py
@@ -1,5 +1,5 @@
 import unittest
-
+import logging
 import rdfframework.connections.blazegraph as blazegraph
 
 class TestBlazegraphDefaultInit(unittest.TestCase):
@@ -8,6 +8,7 @@ class TestBlazegraphDefaultInit(unittest.TestCase):
         pass
 
     def test_no_params(self):
+        blazegraph.Blazegraph.log_level = logging.CRITICAL
         self.assertTrue(blazegraph.Blazegraph().ext_url is not None)
 
     def tearDown(self):
